Create parser module
